requirements.txt
Authlib==1.3.0
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.6
idna==3.7
kaggle==1.6.14
numpy==1.26.4
pandas==2.2.2
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
six==1.16.0
text-unidecode==1.3
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
